backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
passlib[sha256_crypt]
python-jose[cryptography]